pyproject.toml
[project]
name = "pos-venda"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.25.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "python-telegram-bot>=21.0.0",
    "apscheduler>=3.10.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["agents"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
